refactor(LottoRepository): 디렉토리 변경
